examples/rl/driving_in_traffic/requirements.txt
absl-py==0.15.0
astunparse==1.6.3
atari-py==0.2.9
attrs==21.2.0
Automat==20.2.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
cloudpickle==1.3.0
colorama==0.4.4
commonmark==0.9.1
constantly==15.1.0
crafter==1.6.3
cycler==0.11.0
decorator==5.1.0
dm-control==0.0.403778684
dm-env==1.5
dm-tree==0.1.6
dreamerv2==2.1.1
evdev==1.4.0
flatbuffers==1.12
fonttools==4.28.2
future==0.18.2
gast==0.3.3
glfw==2.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
gym==0.18.3
h5py==2.10.0
hyperlink==21.0.0
idna==3.3
ijson==3.1.4
imageio==2.9.0
importlib-metadata==4.8.2
incremental==21.3.0
joblib==1.1.0
jsonpatch==1.32
jsonpointer==2.2
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
labmaze==1.0.5
lxml==4.6.4
Markdown==3.3.6
matplotlib==3.5.0
networkx==2.6.3
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.4.60
opensimplex==0.3
opt-einsum==3.3.0
packaging==21.3
Panda3D==1.10.9
panda3d-gltf==0.13
panda3d-simplepbr==0.9
pandas==1.3.4
Pillow==8.2.0
protobuf==3.19.1
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybullet==3.0.6
pyglet==1.5.15
Pygments==2.10.0
pynput==1.7.5
PyOpenGL==3.1.5
pyparsing==3.0.6
python-dateutil==2.8.2
python-xlib==0.31
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
rich==10.15.1
rsa==4.8
Rtree==0.9.7
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
scikit-image==0.18.3
scikit-learn==1.0.1
scipy==1.7.3
setuptools-scm==6.3.2
sh==1.14.2
Shapely==1.8.0
six==1.15.0
sklearn==0.0
tableprint==0.9.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
tensorflow-probability==0.12.2
termcolor==1.1.0
threadpoolctl==3.0.0
tifffile==2021.11.2
tomli==1.2.2
torchfile==0.1.0
tornado==6.1
tqdm==4.62.3
trimesh==3.9.29
Twisted==21.7.0
typing-extensions==3.7.4.3
urllib3==1.26.7
visdom==0.1.8.9
wcwidth==0.2.5
websocket-client==1.2.1
Werkzeug==2.0.2
wrapt==1.12.1
yattag==1.14.0
zipp==3.6.0
zope.interface==5.4.0